Started to decouple and simplify source and downstream operations.
